Refactoring project's structure. Creating different files for each of the processes. Changes on graphics file to move the chicken and draw the sprites.
