+++ CMakeLists.txt
@@ -5,7 +5,7 @@ INCLUDE_DIRECTORIES(/usr/lib/x86_64-linux-gnu)
 
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11 -pthread -lm -lallegro")
-set(SOURCE_FILES main.cpp utils/utils.cpp utils/graphics.cpp)
+set(SOURCE_FILES main.cpp src/mathfunc.cpp src/graphics.cpp include/globals.h src/bot.cpp include/bot.h src/chicken.cpp include/chicken.h src/nivlem.cpp include/nivlem.h src/workflow.cpp include/workflow.h src/utils.cpp include/utils.h include/externs.h)
 add_executable(NivlemHenHouse  ${SOURCE_FILES})
 
 TARGET_LINK_LIBRARIES(NivlemHenHouse
