+++ include/bot.h
@@ -1,0 +1,12 @@
+//
+// Created by melalonso on 11/10/15.
+//
+
+#ifndef NIVLEMHENHOUSE_BOT_H
+#define NIVLEMHENHOUSE_BOT_H
+
+void *bot_function(void*);
+void *check_food(void*);
+void *check_water(void*);
+
+#endif //NIVLEMHENHOUSE_BOT_H
