+++ include/chicken.h
@@ -1,0 +1,13 @@
+//
+// Created by melalonso on 11/10/15.
+//
+
+#ifndef NIVLEMHENHOUSE_CHICKEN_H
+#define NIVLEMHENHOUSE_CHICKEN_H
+
+void *eat(void *args);
+void *drink(void *args);
+void *swot(void *args);
+void *chicken_process(void * args);
+
+#endif //NIVLEMHENHOUSE_CHICKEN_H
