+++ include/utils.h
@@ -1,0 +1,10 @@
+//
+// Created by melalonso on 11/11/15.
+//
+
+#ifndef NIVLEMHENHOUSE_UTILS_H
+#define NIVLEMHENHOUSE_UTILS_H
+
+void read_file();
+
+#endif //NIVLEMHENHOUSE_UTILS_H
