+++ include/workflow.h
@@ -1,0 +1,12 @@
+//
+// Created by melalonso on 11/10/15.
+//
+
+#ifndef NIVLEMHENHOUSE_WORKFLOW_H
+#define NIVLEMHENHOUSE_WORKFLOW_H
+
+void create_chickens();
+void *count_days(void*);
+void *check_simulation_end(void*);
+
+#endif //NIVLEMHENHOUSE_WORKFLOW_H
